continue code cleanup
